Fixed - FILTER_SANITIZE_STRING deprecated
This code defines a switch statement that sanitizes user input based on the type of filter specified. If the filter is set to FILTER_SANITIZE_FULL_SPECIAL_CHARS, the input is sanitized using filter_input with the specified options, and then any HTML tags are removed using preg_replace. If the filter is set to anything else, the input is sanitized using filter_input without any options, and then any HTML tags are removed using preg_replace.